+++ .storybook/config.js
@@ -1,0 +1,9 @@
+import { configure } from '@storybook/react'
+
+// automatically import all files ending in *.stories.js
+const req = require.context('../src', true, /.story.js$/)
+function loadStories() {
+    req.keys().forEach(filename => req(filename))
+}
+
+configure(loadStories, module)
