+++ package.json
@@ -1,94 +1,103 @@
 {
-    "name": "awesome-talks",
-    "description": "Awesome Talks",
-    "license": "MIT",
-    "version": "0.0.1",
-    "dependencies": {
-        "@fortawesome/fontawesome": "1.1.8",
-        "@fortawesome/fontawesome-free-solid": "5.0.13",
-        "@fortawesome/react-fontawesome": "0.0.20",
-        "apollo-boost": "^0.1.6",
-        "apollo-cache-inmemory": "^1.2.1",
-        "apollo-client": "^2.3.1",
-        "apollo-link": "^1.2.2",
-        "apollo-link-http": "^1.5.4",
-        "apollo-link-state": "^0.4.1",
-        "card-vibes": "0.0.3",
-        "eslint-config-react-app": "^2.1.0",
-        "express": "^4.16.3",
-        "formik": "^0.11.11",
-        "graphql": "0.13.2",
-        "graphql-tag": "^2.9.2",
-        "history": "^4.7.2",
-        "isomorphic-fetch": "^2.2.1",
-        "lodash.debounce": "4.0.8",
-        "random-id": "^0.0.2",
-        "razzle": "^2.0.0-alpha.12",
-        "react": "^16.4.0",
-        "react-apollo": "2.1.4",
-        "react-dom": "^16.4.0",
-        "react-helmet": "^5.2.0",
-        "react-modal": "^3.4.4",
-        "react-router-dom": "^4.2.2",
-        "react-styled-flexboxgrid": "2.2.0",
-        "react-youtube": "^7.6.0",
-        "remcalc": "^1.0.10",
-        "shuffle-array": "^1.0.1",
-        "styled-components": "3.2.6",
-        "styled-flex-component": "2.2.2",
-        "styled-is": "1.1.3",
-        "url-parse": "^1.4.0"
-    },
-    "devDependencies": {
-        "all-contributors-cli": "^4.11.1",
-        "babel-eslint": "^7.2.3",
-        "babel-plugin-auto-import-react": "^4.0.0",
-        "babel-plugin-lodash": "^3.3.2",
-        "babel-plugin-styled-components": "^1.5.1",
-        "babel-plugin-transform-class-properties": "^6.24.1",
-        "babel-plugin-transform-object-rest-spread": "^6.26.0",
-        "babel-plugin-transform-runtime": "^6.23.0",
-        "babel-preset-env": "^1.7.0",
-        "babel-preset-razzle": "^2.0.0-alpha.12",
-        "babel-preset-react": "^6.24.1",
-        "cypress": "^2.1.0",
-        "eslint": "^4.1.1",
-        "eslint-config-standard": "11.0.0-beta.0",
-        "eslint-plugin-flowtype": "^2.34.1",
-        "eslint-plugin-import": "^2.6.0",
-        "eslint-plugin-jsx-a11y": "^5.1.1",
-        "eslint-plugin-node": "^6.0.1",
-        "eslint-plugin-promise": "^3.7.0",
-        "eslint-plugin-react": "^7.1.0",
-        "eslint-plugin-standard": "^3.1.0",
-        "husky": "^0.14.3",
-        "if-env": "^1.0.0",
-        "lint-staged": "^7.1.1",
-        "lodash-webpack-plugin": "^0.11.5",
-        "offline-plugin": "^5.0.3",
-        "prettier": "^1.12.1",
-        "webpack-bundle-analyzer": "^2.13.1"
-    },
-    "scripts": {
-        "posttest": "npm run format",
-        "start": "NODE_ENV=production node build/server.js",
-        "dev": "razzle start",
-        "build": "razzle build",
-        "format": "prettier --write '**/*.{js,css,md}'",
-        "lint": "eslint . --cache --fix",
-        "test": "npm run lint",
-        "precommit": "lint-staged",
-        "contributors:add": "all-contributors add",
-        "contributors:generate": "all-contributors generate"
-    },
-    "lint-staged": {
-        "*.{js,css,md}": [
-            "prettier --write",
-            "git add"
-        ],
-        "*.js": [
-            "eslint --fix",
-            "git add"
-        ]
-    }
+  "name": "awesome-talks",
+  "description": "Awesome Talks",
+  "license": "MIT",
+  "version": "0.0.1",
+  "dependencies": {
+    "@fortawesome/fontawesome": "1.1.8",
+    "@fortawesome/fontawesome-free-solid": "5.0.13",
+    "@fortawesome/react-fontawesome": "0.0.20",
+    "apollo-boost": "^0.1.6",
+    "apollo-cache-inmemory": "^1.2.1",
+    "apollo-client": "^2.3.1",
+    "apollo-link": "^1.2.2",
+    "apollo-link-http": "^1.5.4",
+    "apollo-link-state": "^0.4.1",
+    "card-vibes": "0.0.3",
+    "eslint-config-react-app": "^2.1.0",
+    "express": "^4.16.3",
+    "formik": "^0.11.11",
+    "graphql": "0.13.2",
+    "graphql-tag": "^2.9.2",
+    "history": "^4.7.2",
+    "isomorphic-fetch": "^2.2.1",
+    "lodash.debounce": "4.0.8",
+    "random-id": "^0.0.2",
+    "razzle": "^2.0.0-alpha.12",
+    "react": "^16.4.0",
+    "react-apollo": "2.1.4",
+    "react-dom": "^16.4.0",
+    "react-helmet": "^5.2.0",
+    "react-modal": "^3.4.4",
+    "react-router-dom": "^4.2.2",
+    "react-styled-flexboxgrid": "2.2.0",
+    "react-youtube": "^7.6.0",
+    "remcalc": "^1.0.10",
+    "shuffle-array": "^1.0.1",
+    "styled-components": "3.2.6",
+    "styled-flex-component": "2.2.2",
+    "styled-is": "1.1.3",
+    "url-parse": "^1.4.0"
+  },
+  "devDependencies": {
+    "all-contributors-cli": "^4.11.1",
+    "babel-eslint": "^7.2.3",
+    "babel-plugin-auto-import-react": "^4.0.0",
+    "babel-plugin-lodash": "^3.3.2",
+    "babel-plugin-styled-components": "^1.5.1",
+    "babel-plugin-transform-class-properties": "^6.24.1",
+    "babel-plugin-transform-object-rest-spread": "^6.26.0",
+    "babel-plugin-transform-runtime": "^6.23.0",
+    "babel-preset-env": "^1.7.0",
+    "babel-preset-razzle": "^2.0.0-alpha.12",
+    "babel-preset-react": "^6.24.1",
+    "cypress": "^2.1.0",
+    "eslint": "^4.1.1",
+    "eslint-config-standard": "11.0.0-beta.0",
+    "eslint-plugin-flowtype": "^2.34.1",
+    "eslint-plugin-import": "^2.6.0",
+    "eslint-plugin-jsx-a11y": "^5.1.1",
+    "eslint-plugin-node": "^6.0.1",
+    "eslint-plugin-promise": "^3.7.0",
+    "eslint-plugin-react": "^7.1.0",
+    "eslint-plugin-standard": "^3.1.0",
+    "husky": "^0.14.3",
+    "if-env": "^1.0.0",
+    "lint-staged": "^7.1.1",
+    "lodash-webpack-plugin": "^0.11.5",
+    "offline-plugin": "^5.0.3",
+    "prettier": "^1.12.1",
+    "sinon": "^5.0.10",
+    "webpack-bundle-analyzer": "^2.13.1",
+    "@storybook/react": "^3.4.6",
+    "@storybook/addon-actions": "^3.4.6",
+    "@storybook/addon-links": "^3.4.6",
+    "@storybook/addons": "^3.4.6",
+    "babel-core": "^6.26.3",
+    "babel-runtime": "^6.26.0"
+  },
+  "scripts": {
+    "posttest": "npm run format",
+    "start": "NODE_ENV=production node build/server.js",
+    "dev": "razzle start",
+    "build": "razzle build",
+    "format": "prettier --write '**/*.{js,css,md}'",
+    "lint": "eslint . --cache --fix",
+    "test": "npm run lint",
+    "precommit": "lint-staged",
+    "contributors:add": "all-contributors add",
+    "contributors:generate": "all-contributors generate",
+    "storybook": "start-storybook -p 6006",
+    "build-storybook": "build-storybook"
+  },
+  "lint-staged": {
+    "*.{js,css,md}": [
+      "prettier --write",
+      "git add"
+    ],
+    "*.js": [
+      "eslint --fix",
+      "git add"
+    ]
+  }
 }
