+++ src/Components/Favorite/story.js
@@ -1,0 +1,13 @@
+import React from 'react'
+
+import { storiesOf } from '@storybook/react'
+
+import { Favorited, UnFavorited } from './example'
+
+storiesOf('Favorite', module)
+    .add('Favorited', () => {
+        return <Favorited />
+    })
+    .add('Unfavorited', () => {
+        return <UnFavorited />
+    })
