+++ src/Components/HideViewed/story.js
@@ -1,0 +1,13 @@
+import React from 'react'
+
+import { storiesOf } from '@storybook/react'
+
+import { Hidden, NotHidden } from './example'
+
+storiesOf('HideViewed', module)
+    .add('Hidden', () => {
+        return <Hidden />
+    })
+    .add('Not Hidden', () => {
+        return <NotHidden />
+    })
